+++ apps/desktop/src-tauri/src/app_commands.rs
@@ -14,7 +14,6 @@ macro_rules! with_app_commands {
             commands::engine_catalog_version => "engine_catalog_version",
             commands::end_popover_hold => "end_popover_hold",
             commands::export_diagnostics => "export_diagnostics",
-            commands::export_session => "export_session",
             commands::finish_onboarding => "finish_onboarding",
             commands::get_consent_diagnostics => "get_consent_diagnostics",
             commands::get_folder_permissions => "get_folder_permissions",